requirements.txt
alembic==1.9.4
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
billiard==3.6.4.0
blinker==1.5
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
email-validator==1.1.3
Flask==2.0.2
Flask-Admin==1.6.1
Flask-Babel==2.0.0
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.0.2
jsmin==3.0.1
jsonschema==4.17.3
kombu==5.2.4
lesscpy==0.15.1
Mako==1.2.4
MarkupSafe==2.1.2
ply==3.11
prompt-toolkit==3.0.38
pypng==0.20220715.0
pyrsistent==0.19.3
python-dotenv==0.19.2
pytz==2022.7.1
qrcode==7.4.2
rcssmin==1.1.1
redis==4.5.1
shortuuid==1.0.11
six==1.16.0
SQLAlchemy==1.4.29
typing_extensions==4.5.0
vine==5.0.0
wcwidth==0.2.6
Werkzeug~=2.0.0
WTForms==3.0.0
xxhash==3.2.0
icalendar
zipp==3.15.0
user-agents
geopy
requests
psycopg2-binary
apscheduler
